requirements.txt
fastapi
uvicorn
python-dotenv
supabase
cloudinary
transformers
torch
Pillow
numpy
python-multipart
cryptography
itsdangerous
python-jose[cryptography]
requests
